+++ common/on-chip-calib.h
@@ -49,7 +49,7 @@ namespace rs2
 
         void update_last_used();
 
-        uint32_t ground_truth = 2500;
+        float ground_truth = 2500;
         int average_step_count = 20;
         int step_count = 20;
         int accuracy = 2;
